+++ jinns/data/__init__.py
@@ -2,7 +2,7 @@ from ._DataGeneratorODE import DataGeneratorODE
 from ._CubicMeshPDEStatio import CubicMeshPDEStatio
 from ._CubicMeshPDENonStatio import CubicMeshPDENonStatio
 from ._DataGeneratorObservations import DataGeneratorObservations
-from ._DataGeneratorParameter import DataGeneratorParameter
+from ._DataGeneratorParameter import DataGeneratorParameter, DGParams
 from ._Batchs import ODEBatch, PDEStatioBatch, PDENonStatioBatch
 
 from ._utils import append_obs_batch, append_param_batch
@@ -12,6 +12,7 @@ __all__ = [
     "CubicMeshPDEStatio",
     "CubicMeshPDENonStatio",
     "DataGeneratorParameter",
+    "DGParams",
     "DataGeneratorObservations",
     "ODEBatch",
     "PDEStatioBatch",