+++ jinns/utils/__init__.py
@@ -1,3 +1,4 @@
 from ._utils import get_grid
+from ._DictToModuleMeta import DictToModuleMeta
 
-__all__ = ["get_grid"]
+__all__ = ["get_grid", "DictToModuleMeta"]