+++ tests/operator_tests/test_laplacian_fwd.py
@@ -23,6 +23,7 @@ def create_u_statio():
 
 def test_laplacian_fwd_statio(create_u_statio):
     u_statio, params = create_u_statio
+    params = jinns.parameters.Params(nn_params=params, eq_params={})
     x = jnp.stack([jnp.array([0.4, 1.5]) for i in range(10)], axis=0)
 
     assert jnp.allclose(
@@ -49,6 +50,7 @@ def create_u_nonstatio():
 
 def test_laplacian_fwd_nonstatio(create_u_nonstatio):
     u_nonstatio, params = create_u_nonstatio
+    params = jinns.parameters.Params(nn_params=params, eq_params={})
     t_x = jnp.stack([jnp.array([0.5, 0.4, 1.5]) for i in range(10)], axis=0)
     assert jnp.allclose(
         jinns.loss.laplacian_fwd(t_x, u_nonstatio, params, method="loop"),